detect-tls-cert-validity: move unittests to tests/
